+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.20.1] Q1 2024
+- added distroless docker image
+- refactored partition handling of file functions
+- print version info on startup
+
 ## [0.20.0] Q1 2024
 - device update import command:
   - splitted import command into create-import-manifest and import-update